+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Ensure `DocumentVisibility` implements `arez.Disposable`
+
 ### [v0.52](https://github.com/arez/arez-dom/tree/v0.52) (2020-02-11) · [Full Changelog](https://github.com/arez/arez-dom/compare/v0.51...v0.52)
 
 * Add `DocumentVisibility.isVisible()` and `DocumentVisibility.isHidden()` helper methods.